order file validation improved

+++ src/FileManager.java
@@ -1,5 +1,6 @@
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
+import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
@@ -51,28 +52,32 @@ public class FileManager {
     ArrayList<Integer> numbers = new ArrayList<Integer>();
     try {
         //  READ
-        BufferedReader reader = new BufferedReader(new FileReader(fileName + ".csv"));
 
-        String line;
-        while ((line = reader.readLine()) != null) {
-            numbers.add(Integer.parseInt((line.split(",")[0])));
+        String filePath = "./"+ fileName + ".csv";
+        File file = new File(filePath);
+        if (file.exists()) {
+            BufferedReader reader = new BufferedReader(new FileReader("./"+ fileName + ".csv"));
+            String line;
+            while ((line = reader.readLine()) != null) {
+                numbers.add(Integer.parseInt((line.split(",")[0])));
+            }
+    
+            reader.close();
+    
+            numbers = (ArrayList<Integer>) MergeSort.sort(numbers);
+
+            //  WRITE
+            String newFileName = fileName.split(".csv")[0] + "-ordenado" + ".csv";
+            BufferedWriter writer = new BufferedWriter(new FileWriter(newFileName));
+
+            for (int i = 0; i < numbers.size(); ++i) {
+                writer.write(numbers.get(i) + "\n");
+            }
+            System.out.println("El archivo '" + fileName + "-ordenado.csv' fue creado.");
+            writer.close();
+        }else {
+            System.out.println("El archivo '" + fileName + ".csv' no existe.");
         }
-
-        reader.close();
-
-        numbers = (ArrayList<Integer>) MergeSort.sort(numbers);
-
-
-        //  WRITE
-        String newFileName = fileName.split(".csv")[0] + "-ordenado" + ".csv";
-        BufferedWriter writer = new BufferedWriter(new FileWriter(newFileName));
-
-        for (int i = 0; i < numbers.size(); ++i) {
-            writer.write(numbers.get(i) + "\n");
-        }
-
-        writer.close();
-
 
     } catch (IOException e) {
         e.printStackTrace();
